Version 1 Complete
This works, it appears to pass all tests. Testing is not exhaustive.
Not all known errors are currently accounted for - for instance, a row might still contain an impossible combination of values such as ADDRESS_START or ADDRESS_END without the prerequisite values ADDRESS_1, ADDRESS_2,....
This can be accounted for but requires more research first.